system-services: use new generic service interface
Previously generic service returned a string, not the raw values
returned by the system controller. This didn't make very much sense -
why would the driver do this when any serious use of it would want the
values themselves rather than a string - ie the signature demo had to
work around this behaviour. Rework the examples to use the raw values
rather and derive the strings themselves.

Signed-off-by: Conor Dooley <conor.dooley@microchip.com>