system-services: explicitly set response offset relative to send offset
The response offset is directly related to the send offset. The send
offset is the address in the mailbox that the system controller should
start reading from (or writing to*). The response offset is offset from
the send offset by at least zero:
- If there is no data being sent to the system controller, then the
  response offset is identical to the send offset.
- If data is sent to the system controller, the response offset can be
  either the send offset *or* the send offset plus the size of the data
  sent. In the former case, the size of the response should be
  incremented by the size of the data sent. If data is sent to the
  system controller the response will be appended to the sent data
  rather than replace it, hence the increase by the send data size.

While this has no functional change, change the examples to tie these
two parameters together.

Signed-off-by: Conor Dooley <conor.dooley@microchip.com>