dma,lsram: update address based on 2022.09 design
When users try to add onto the reference design, they find that the
current addresses that peripherals connected to Fabric InterConnect
(FIC) 3 use are restrictive. For the v2022.09 reference design, the
peripherals have been shifted down, leaving more contiguous address
space for their custom IP/peripherals.

Update the lsram and dma examples with the new addresses used in the
2022.09 reference design.

Signed-off-by: Lewis Hanly <lewis.hanly@microchip.com>