dt-overlays: update readme due to overlay changes
The Raspberry Pi Sense Hat dtso is updated for the 6.1 kernel. With this update
node names have been changed and the sensehat core module has been removed. The
MFD_SIMPLE_MFD_I2C module replaces the sensehat core one and this new module is
selected with the INPUT_JOYSTICK config. This means the modprobe command is no
longer necessary to load the module. Remove the modprobe command and update
pressure sensor references

Signed-off-by: Jamie Gibbons <jamie.gibbons@microchip.com>