Industrial Edge Demo: adding readme for demos
Adding the README file link, and how to run the demo for Industrial
Edge Stack. This design describes how to run the OPC-UA Industrial
Edge Demo and demonstrates the basic functionality of the PolarFire
SoC Video Kit, the PolarFire SoC ICICLE Kit, the Stepper Motor, and
their communication with each other over the OPC-UA protocol.

Signed-off-by: Ahmed Mujtaba Ali <ahmed.ali2@microchip.com>