can: drop uio can example and add socketCAN documentation instead
Starting from 2024.02 release, the UIO CAN driver has been dropped in
favour of the socketCAN interface, which allows to use the CAN interface
with standard Linux tools such as can-utils and iproute.

Drop the CAN UIO example and update the README in the "can" directory
to provide information on how to configure and test the CAN interface(s)
on PolarFire SoC using the iproute tool.

Signed-off-by: Valentina Fernandez <valentina.fernandezalanis@microchip.com>