iio-http-server: update to mainline userspace api
When the driver was mainlined, a rake of changes were made to the
userspace API. The 6.6 version of Linux4Microchip includes a backport of
the mainline driver, so update the example to use the new paths.

Voltage scale is now per channel, so adjust the code to account for that
also.

Signed-off-by: Conor Dooley <conor.dooley@microchip.com>