Add script to update gateware on the BeagleV-Fire
The BeagleV-Fire gateware is a bitstream file that can contain components like the FPGA design, a Microprocessor Subsystem (MSS) configuration,
and a set of device tree overlays describes the contents of the FPGA fabric.

Add a script which is used by the BeagleV-Fire to update the gateware, allowing customization of the BeagleV-Fire connectors(CAPES) and/or
updates to the FPGA design using the auto-update framework.

Signed-off-by: prashanth kumar burujukindi <prashanthkumar.burujukindi@microchip.com>