gateware: update gateware script
Update flash_erase to only erase dtbos as erasing such a large amount
of memory was overkill. A reboot, and therefore a firmware update is
only desired if no error has occured in the validation stage. Add an
additional conditional to only reboot if no error is reported. Use cat
instead of dd command to set-up data as suggested by auto-update
documentation. Update formatting and comments to match script in
Buildroot - keeping a uniform copy for all build systems.

Signed-off-by: Jamie Gibbons <jamie.gibbons@microchip.com>