gateware: support passing gateware/dtbo files explicitly
On production kits, it may not be the case where a dtbo file is
provided, fbe that not using the same infrastructure as the beagle
to produce it automatically or something else. Permit passing the
gateware and dtbo files explicitly, rather than just the directory
containing files matching the default naming convention. This has the
additional benefit of making keeping track of what is being programmed
easier in an setting where there are multiple images involved.

Copying files into /lib/firmware/ didn't seem like a required part of
the process, so I removed it as part of this change.

Signed-off-by: Conor Dooley <conor.dooley@microchip.com>