gateware: drop mpfs_dtbo.spi bits from buildroot section
The document linked in the README does not explain how to generate
an mpfs_dtbo.fpi file, containing the design info but uses commands that
expect there to be one. Rather than explain how to generate one, which
would require a fresh document written, switch to commands and
instructions that are intended to program only a .spi bitstream file.

Signed-off-by: Conor Dooley <conor.dooley@microchip.com>